requirements.txt
Django==1.8
django-endless-pagination==2.0
django-likes==0.2
django-multiuploader==0.2.9
django-passwords==0.3.7
django-secretballot==0.5.0
django-tastypie==0.13.3
django-widget-tweaks==1.4.1
Pillow==3.1.1
python-dateutil==2.5.0
python-magic==0.4.10
python-mimeparse==1.5.1
six==1.10.0
sorl-thumbnail==12.3
wheel==0.29.0
